# server/swagger.yaml
openapi: 3.0.3
info:
  title: Limble Pie Chart API
  description: |-
    This is a sample Pie Chart API
  version: 0.0.1
servers:
  - url: http://127.0.0.1:3000
...
paths:
...
  /tasks/worker:
    post:
      tags:
        - pie
      summary: Get tasks grouped by workers
      description: ''
      operationId: tasksWorkers
      requestBody:
        description: Params to filter tasks
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PieRequest'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PieResponse'
        '400':
          description: Invalid input
components:
  schemas:
    PieRequest:
      type: object
      properties:
        is_complete:
          type: boolean
          example: true|false
        location_ids:
          type: array
          items:
            type: integer
          example: [1,2,3,4]
        worker_ids:
          type: array
          items:
            type: integer
          example: [1,2,3,4]
...
    PieResponse:
      type: object
      properties:
        result:
          type: string
          example: [OK|ERROR]
        tasks:
          type: array
          items:
            type: object
            properties:
              cost:
                type: number
                example: 32.777777
              hourly_wage:
                type: number
                example: 10.00
              id:
                type: integer
              location_ids:
                type: string
              locations_cnt:
                type: integer
              logged_time_ids:
                type: string
              task_ids:
                type: string
              tasks_cnt:
                type: integer
              username:
                type: string
              worker_ids:
                type: string
...